# Backend/github-discovery/output/20edbcb95ceffb10_openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "API Specification - 20edbcb95ceffb10",
    "version": "1.0.0",
    "description": "Auto-generated from repository https://github.com/ilyeschaabani/microservice.git",
    "x-generated": "2026-03-23T16:12:12.162345Z",
    "x-endpoint-count": 34
  },
  "servers": [
    {
      "url": "https://api.example.com"
    }
  ],
  "paths": {
...
    "/evaluations/evaluations/add": {
      "post": {
        "summary": "createEvaluation",
        "tags": [
          "create",
          "evaluation"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "x-metadata": {
          "annotation": "@PostMapping(\"/add\""
        },
        "x-confidence": 0.87,
        "x-source": "ast"
      }
    },
    "/evaluations/evaluations/listeval": {
      "get": {
        "summary": "getAllEvaluations",
        "tags": [
          "get",
          "evaluation"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "x-metadata": {
          "annotation": "@GetMapping(\"/listeval\""
        },
        "x-confidence": 0.87,
        "x-source": "ast"
      }
    },
    "/evaluations/evaluations/{idEvaluation}": {
      "put": {
        "summary": "updateEvaluation",
        "tags": [
          "update",
          "evaluation"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "x-metadata": {
          "annotation": "@PutMapping(\"/{idEvaluation}\""
        },
        "x-confidence": 0.87,
        "x-source": "ast"
      },
      "delete": {
        "summary": "deleteEvaluation",
        "tags": [
          "delete",
          "evaluation"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "x-metadata": {
          "annotation": "@DeleteMapping(\"/{idEvaluation}\""
        },
        "x-confidence": 0.87,
        "x-source": "ast"
      },
      "get": {
        "summary": "getEvaluationById",
        "tags": [
          "get",
          "evaluation"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "x-metadata": {
          "annotation": "@GetMapping(\"/{idEvaluation}\""
        },
        "x-confidence": 0.87,
        "x-source": "ast"
      }
    },
    "/evaluations/evaluations/{idEvaluation}/questions": {
      "post": {
        "summary": "addQuestionsToEvaluation",
        "tags": [
          "add",
          "evaluation"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "x-metadata": {
          "annotation": "@PostMapping(\"/{idEvaluation}/questions\""
        },
        "x-confidence": 0.87,
        "x-source": "ast"
      }
    },
    "/questions/questions/{quesId}": {
      "put": {
        "summary": "updateQuestion",
        "tags": [
          "update",
          "question"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "x-metadata": {
          "annotation": "@PutMapping(\"/{quesId}\""
        },
        "x-confidence": 0.87,
        "x-source": "ast"
      },
      "delete": {
        "summary": "deleteQuestion",
        "tags": [
          "question",
          "delete"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "x-metadata": {
          "annotation": "@DeleteMapping(\"/{quesId}\""
        },
        "x-confidence": 0.87,
        "x-source": "ast"
      }
    },
    "/questions/questions/by-evaluation/{idEvaluation}": {
      "get": {
        "summary": "getQuestionsByEvaluation",
        "tags": [
          "get",
          "question"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "x-metadata": {
          "annotation": "@GetMapping(\"/by-evaluation/{idEvaluation}\""
        },
        "x-confidence": 0.87,
        "x-source": "ast"
      }
    },
    "/api/quiz/api/quiz/questions": {
      "get": {
        "summary": "getQuestions",
        "tags": [
          "get",
          "quiz"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "x-metadata": {
          "annotation": "@GetMapping(\"/questions\""
        },
        "x-confidence": 0.87,
        "x-source": "ast"
      }
    },
    "/api/quiz/api/quiz/evaluate": {
      "post": {
        "summary": "evaluateQuiz",
        "tags": [
          "evaluate",
          "quiz"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "x-metadata": {
          "annotation": "@PostMapping(\"/evaluate\""
        },
        "x-confidence": 0.87,
        "x-source": "ast"
      }
    },
...
    "/api/ressources/api/ressources/{id}": {
      "get": {
        "summary": "getRessourceById",
        "tags": [
          "get",
          "ressources"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "x-metadata": {
          "annotation": "@GetMapping(\"/{id}\""
        },
        "x-confidence": 0.87,
        "x-source": "ast"
      },
      "put": {
        "summary": "updateRessource",
        "tags": [
          "update",
          "ressources"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "x-metadata": {
          "annotation": "@PutMapping(\"/{id}\""
        },
        "x-confidence": 0.87,
        "x-source": "ast"
      },
      "delete": {
        "summary": "deleteRessource",
        "tags": [
          "ressources",
          "delete"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "x-metadata": {
          "annotation": "@DeleteMapping(\"/{id}\""
        },
        "x-confidence": 0.87,
        "x-source": "ast"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
}